Ordenamiento de funciones
Ordene las funciones para una mejor claridad(me estaba volviendo loco)
