+++ website/templates/contenido.html
@@ -148,10 +148,8 @@
 
     </footer>
 
-    <script type="module" src="{{ url_for('static', filename='js/mostrar_datos.js') }}"></script>
+    <script type="module" src="{{ url_for('static', filename='js/main.js') }}"></script>
     <script src="{{ url_for('static', filename='js/perfil.js') }}"></script>
-
-    <!-- <script src="{{ url_for('static', filename='js/carrusel.js') }}"></script> -->
     
 
 </body>
